+++ apps/web/src/app/[locale]/landing/RankingSection/SvgContainer.tsx
@@ -2,235 +2,218 @@
 
 import { css } from '_panda/css';
 
-export function GameConsoleSvgContainer({
-  children,
-  joystickRotation,
-}: {
-  children: React.ReactNode;
-  joystickRotation: number;
-}) {
+export function GameConsoleSvgContainer({ children }: { children: React.ReactNode }) {
   return (
-    <div className={containerStyle}>
-      <div className={svgContainerStyle}>
-        <svg
-          viewBox="0 0 1660 1714"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-          className={svgStyle}
-          preserveAspectRatio="xMidYMid meet"
-        >
-          <rect
-            x="151"
-            y="10"
-            width="1358"
-            height="1212"
-            fill="url(#paint0_linear_2620_81262)"
-            stroke="url(#paint1_linear_2620_81262)"
-            strokeWidth="20"
-          />
-          <path
-            d="M17.3404 1474L146.12 1222H1511.94L1643.5 1474H17.3404Z"
-            fill="url(#paint2_linear_2620_81262)"
-            stroke="url(#paint3_linear_2620_81262)"
-            strokeWidth="20"
-          />
+    <div className={svgContainerStyle}>
+      <svg
+        viewBox="0 0 1660 1714"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+        className={svgStyle}
+        preserveAspectRatio="xMidYMid meet"
+      >
+        <rect
+          x="151"
+          y="10"
+          width="1358"
+          height="1212"
+          fill="url(#paint0_linear_2620_81262)"
+          stroke="url(#paint1_linear_2620_81262)"
+          strokeWidth="20"
+        />
+        <path
+          d="M17.3404 1474L146.12 1222H1511.94L1643.5 1474H17.3404Z"
+          fill="url(#paint2_linear_2620_81262)"
+          stroke="url(#paint3_linear_2620_81262)"
+          strokeWidth="20"
+        />
 
-          {/* Power Button - Interactive */}
+        {/* Power Button - Interactive */}
 
-          <foreignObject x="187" y="33" width="1286" height="1148">
-            <div className={blurContainerStyle} />
-          </foreignObject>
+        <foreignObject x="187" y="33" width="1286" height="1148">
+          <div className={blurContainerStyle} />
+        </foreignObject>
 
-          <path
-            data-figma-bg-blur-radius="40"
-            d="M237 78H232V83V1131V1136H237H1423H1428V1131V83V78H1423H237Z"
-            fill="#6DE575"
-            stroke="#A1FFA7"
-            strokeWidth="10"
-          />
+        <path
+          data-figma-bg-blur-radius="40"
+          d="M237 78H232V83V1131V1136H237H1423H1428V1131V83V78H1423H237Z"
+          fill="#6DE575"
+          stroke="#A1FFA7"
+          strokeWidth="10"
+        />
 
-          {/* Screen Background */}
-          <path d="M280 121H1380V1086H280V121Z" fill="url(#paint4_linear_2620_81262)" />
+        {/* Screen Background */}
+        <path d="M280 121H1380V1086H280V121Z" fill="url(#paint4_linear_2620_81262)" />
 
-          {children}
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M1320 121H1300V141H1320V161H1340V181H1360V201H1380V181V161V141V126V121H1370H1360H1340H1320Z"
-            fill="#6EE577"
-          />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M1320 1086H1300V1066H1320V1046H1340V1026H1360V1006H1380V1026V1046V1066V1081V1086H1370H1360H1340H1320Z"
-            fill="#6EE577"
-          />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M340 1086H360V1066H340V1046H320V1026H300V1006H280V1026V1046V1066V1081V1086H290H300H320H340Z"
-            fill="#6EE577"
-          />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M340 121H360V141H340V161H320V181H300V201H280V181V161V141V121H300H320H340Z"
-            fill="#6EE577"
-          />
-          <rect x="1340" y="121" width="20" height="20" fill="#00C448" />
-          <rect x="1360" y="141" width="20" height="20" fill="#00C448" />
-          <rect width="20" height="20" transform="matrix(1 0 0 -1 1340 1086)" fill="#00C448" />
-          <rect width="20" height="20" transform="matrix(1 0 0 -1 1360 1066)" fill="#00C448" />
-          <rect x="320" y="1086" width="20" height="20" transform="rotate(180 320 1086)" fill="#00C448" />
-          <rect x="300" y="1066" width="20" height="20" transform="rotate(180 300 1066)" fill="#00C448" />
-          <rect width="20" height="20" transform="matrix(-1 0 0 1 320 121)" fill="#00C448" />
-          <rect width="20" height="20" transform="matrix(-1 0 0 1 300 141)" fill="#00C448" />
-          <path
-            d="M28 1474H1632C1641.94 1474 1650 1482.06 1650 1492V1704H10V1492C10 1482.06 18.0589 1474 28 1474Z"
-            fill="url(#paint5_linear_2620_81262)"
-            stroke="#0799D7"
-            strokeWidth="20"
-          />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M1479 0H1499H1507H1519V13V20V40H1499V20H1479V0Z"
-            fill="#141517"
-          />
-          <rect x="1479" y="20" width="20" height="20" fill="#0799D7" />
-          <path fillRule="evenodd" clipRule="evenodd" d="M181 0H161H153H141V13V20V40H161V20H181V0Z" fill="#141517" />
-          <rect width="20" height="20" transform="matrix(-1 0 0 1 181 20)" fill="#0799D7" />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M1394 73H1414H1421H1434V86V93V113H1414V93H1394V73Z"
-            fill="#0477D9"
-          />
-          <rect x="1384" y="80" width="10" height="16" fill="#A0FFA7" />
-          <rect x="1404" y="100" width="10" height="16" fill="#A0FFA7" />
-          <rect x="1384" y="93" width="15" height="10" fill="#A0FFA7" />
-          <rect x="1394" y="93" width="20" height="10" fill="#A0FFA7" />
-          <rect x="1404" y="113" width="23" height="10" fill="#A0FFA7" />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M1393 1141H1413H1420H1433V1128V1121V1101H1413V1121H1393V1141Z"
-            fill="#51C779"
-          />
-          <rect width="10" height="16" transform="matrix(1 0 0 -1 1383 1134)" fill="#A0FFA7" />
-          <rect width="10" height="16" transform="matrix(1 0 0 -1 1403 1114)" fill="#A0FFA7" />
-          <rect width="15" height="10" transform="matrix(1 0 0 -1 1383 1121)" fill="#A0FFA7" />
-          <rect width="20" height="10" transform="matrix(1 0 0 -1 1393 1121)" fill="#A0FFA7" />
-          <rect width="23" height="10" transform="matrix(1 0 0 -1 1403 1101)" fill="#A0FFA7" />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M267 1141H247H240H227V1128V1121V1101H247V1121H267V1141Z"
-            fill="#51C779"
-          />
-          <rect x="277" y="1134" width="10" height="16" transform="rotate(180 277 1134)" fill="#A0FFA7" />
-          <rect x="257" y="1114" width="10" height="16" transform="rotate(180 257 1114)" fill="#A0FFA7" />
-          <rect x="277" y="1121" width="15" height="10" transform="rotate(180 277 1121)" fill="#A0FFA7" />
-          <rect x="267" y="1121" width="20" height="10" transform="rotate(180 267 1121)" fill="#A0FFA7" />
-          <rect x="257" y="1101" width="23" height="10" transform="rotate(180 257 1101)" fill="#A0FFA7" />
-          <path fillRule="evenodd" clipRule="evenodd" d="M267 73H247H240H227V86V93V113H247V93H267V73Z" fill="#0477D9" />
-          <rect width="10" height="16" transform="matrix(-1 0 0 1 277 80)" fill="#A0FFA7" />
-          <rect width="10" height="16" transform="matrix(-1 0 0 1 257 100)" fill="#A0FFA7" />
-          <rect width="15" height="10" transform="matrix(-1 0 0 1 277 93)" fill="#A0FFA7" />
-          <rect width="20" height="10" transform="matrix(-1 0 0 1 267 93)" fill="#A0FFA7" />
-          <rect width="23" height="10" transform="matrix(-1 0 0 1 257 113)" fill="#A0FFA7" />
-          <rect x="276" y="1348" width="182" height="40" fill="#1C2923" />
+        {children}
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M1320 121H1300V141H1320V161H1340V181H1360V201H1380V181V161V141V126V121H1370H1360H1340H1320Z"
+          fill="#6EE577"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M1320 1086H1300V1066H1320V1046H1340V1026H1360V1006H1380V1026V1046V1066V1081V1086H1370H1360H1340H1320Z"
+          fill="#6EE577"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M340 1086H360V1066H340V1046H320V1026H300V1006H280V1026V1046V1066V1081V1086H290H300H320H340Z"
+          fill="#6EE577"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M340 121H360V141H340V161H320V181H300V201H280V181V161V141V121H300H320H340Z"
+          fill="#6EE577"
+        />
+        <rect x="1340" y="121" width="20" height="20" fill="#00C448" />
+        <rect x="1360" y="141" width="20" height="20" fill="#00C448" />
+        <rect width="20" height="20" transform="matrix(1 0 0 -1 1340 1086)" fill="#00C448" />
+        <rect width="20" height="20" transform="matrix(1 0 0 -1 1360 1066)" fill="#00C448" />
+        <rect x="320" y="1086" width="20" height="20" transform="rotate(180 320 1086)" fill="#00C448" />
+        <rect x="300" y="1066" width="20" height="20" transform="rotate(180 300 1066)" fill="#00C448" />
+        <rect width="20" height="20" transform="matrix(-1 0 0 1 320 121)" fill="#00C448" />
+        <rect width="20" height="20" transform="matrix(-1 0 0 1 300 141)" fill="#00C448" />
+        <path
+          d="M28 1474H1632C1641.94 1474 1650 1482.06 1650 1492V1704H10V1492C10 1482.06 18.0589 1474 28 1474Z"
+          fill="url(#paint5_linear_2620_81262)"
+          stroke="#0799D7"
+          strokeWidth="20"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M1479 0H1499H1507H1519V13V20V40H1499V20H1479V0Z"
+          fill="#141517"
+        />
+        <rect x="1479" y="20" width="20" height="20" fill="#0799D7" />
+        <path fillRule="evenodd" clipRule="evenodd" d="M181 0H161H153H141V13V20V40H161V20H181V0Z" fill="#141517" />
+        <rect width="20" height="20" transform="matrix(-1 0 0 1 181 20)" fill="#0799D7" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M1394 73H1414H1421H1434V86V93V113H1414V93H1394V73Z"
+          fill="#0477D9"
+        />
+        <rect x="1384" y="80" width="10" height="16" fill="#A0FFA7" />
+        <rect x="1404" y="100" width="10" height="16" fill="#A0FFA7" />
+        <rect x="1384" y="93" width="15" height="10" fill="#A0FFA7" />
+        <rect x="1394" y="93" width="20" height="10" fill="#A0FFA7" />
+        <rect x="1404" y="113" width="23" height="10" fill="#A0FFA7" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M1393 1141H1413H1420H1433V1128V1121V1101H1413V1121H1393V1141Z"
+          fill="#51C779"
+        />
+        <rect width="10" height="16" transform="matrix(1 0 0 -1 1383 1134)" fill="#A0FFA7" />
+        <rect width="10" height="16" transform="matrix(1 0 0 -1 1403 1114)" fill="#A0FFA7" />
+        <rect width="15" height="10" transform="matrix(1 0 0 -1 1383 1121)" fill="#A0FFA7" />
+        <rect width="20" height="10" transform="matrix(1 0 0 -1 1393 1121)" fill="#A0FFA7" />
+        <rect width="23" height="10" transform="matrix(1 0 0 -1 1403 1101)" fill="#A0FFA7" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M267 1141H247H240H227V1128V1121V1101H247V1121H267V1141Z"
+          fill="#51C779"
+        />
+        <rect x="277" y="1134" width="10" height="16" transform="rotate(180 277 1134)" fill="#A0FFA7" />
+        <rect x="257" y="1114" width="10" height="16" transform="rotate(180 257 1114)" fill="#A0FFA7" />
+        <rect x="277" y="1121" width="15" height="10" transform="rotate(180 277 1121)" fill="#A0FFA7" />
+        <rect x="267" y="1121" width="20" height="10" transform="rotate(180 267 1121)" fill="#A0FFA7" />
+        <rect x="257" y="1101" width="23" height="10" transform="rotate(180 257 1101)" fill="#A0FFA7" />
+        <path fillRule="evenodd" clipRule="evenodd" d="M267 73H247H240H227V86V93V113H247V93H267V73Z" fill="#0477D9" />
+        <rect width="10" height="16" transform="matrix(-1 0 0 1 277 80)" fill="#A0FFA7" />
+        <rect width="10" height="16" transform="matrix(-1 0 0 1 257 100)" fill="#A0FFA7" />
+        <rect width="15" height="10" transform="matrix(-1 0 0 1 277 93)" fill="#A0FFA7" />
+        <rect width="20" height="10" transform="matrix(-1 0 0 1 267 93)" fill="#A0FFA7" />
+        <rect width="23" height="10" transform="matrix(-1 0 0 1 257 113)" fill="#A0FFA7" />
+        <rect x="276" y="1348" width="182" height="40" fill="#1C2923" />
 
-          <defs>
-            <clipPath id="bgblur_0_2620_81262_clip_path" transform="translate(-187 -33)">
-              <path d="M237 78H232V83V1131V1136H237H1423H1428V1131V83V78H1423H237Z" />
-            </clipPath>
-            <linearGradient
-              id="paint0_linear_2620_81262"
-              x1="830"
-              y1="0"
-              x2="830"
-              y2="1232"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stopColor="#016EDB" />
-              <stop offset="0.555" stopColor="#16B7CD" />
-              <stop offset="1" stopColor="#5CCA69" />
-            </linearGradient>
-            <linearGradient
-              id="paint1_linear_2620_81262"
-              x1="830"
-              y1="0"
-              x2="830"
-              y2="1232"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stopColor="#0799D7" />
-              <stop offset="1" stopColor="#00C448" />
-            </linearGradient>
-            <linearGradient
-              id="paint2_linear_2620_81262"
-              x1="830.5"
-              y1="1484"
-              x2="830.5"
-              y2="1212"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stopColor="#AFFFB9" />
-              <stop offset="0.405" stopColor="#9CF3A8" />
-              <stop offset="1" stopColor="#58C96E" />
-            </linearGradient>
-            <linearGradient
-              id="paint3_linear_2620_81262"
-              x1="830.5"
-              y1="1212"
-              x2="830.5"
-              y2="1484"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stopColor="#01C34B" />
-              <stop offset="1" stopColor="#0799D7" />
-            </linearGradient>
-            <linearGradient
-              id="paint4_linear_2620_81262"
-              x1="830"
-              y1="121"
-              x2="830"
-              y2="1086"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop />
-              <stop offset="0.165" stopColor="#001420" />
-              <stop offset="0.365" stopColor="#002943" />
-              <stop offset="1" stopColor="#008FE8" />
-            </linearGradient>
-            <linearGradient
-              id="paint5_linear_2620_81262"
-              x1="830"
-              y1="1464"
-              x2="830"
-              y2="1714"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stopColor="#016EDB" />
-              <stop offset="0.555" stopColor="#16B7CD" />
-              <stop offset="1" stopColor="#5CCA69" />
-            </linearGradient>
-          </defs>
-        </svg>
-      </div>
+        <defs>
+          <clipPath id="bgblur_0_2620_81262_clip_path" transform="translate(-187 -33)">
+            <path d="M237 78H232V83V1131V1136H237H1423H1428V1131V83V78H1423H237Z" />
+          </clipPath>
+          <linearGradient
+            id="paint0_linear_2620_81262"
+            x1="830"
+            y1="0"
+            x2="830"
+            y2="1232"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#016EDB" />
+            <stop offset="0.555" stopColor="#16B7CD" />
+            <stop offset="1" stopColor="#5CCA69" />
+          </linearGradient>
+          <linearGradient
+            id="paint1_linear_2620_81262"
+            x1="830"
+            y1="0"
+            x2="830"
+            y2="1232"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#0799D7" />
+            <stop offset="1" stopColor="#00C448" />
+          </linearGradient>
+          <linearGradient
+            id="paint2_linear_2620_81262"
+            x1="830.5"
+            y1="1484"
+            x2="830.5"
+            y2="1212"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#AFFFB9" />
+            <stop offset="0.405" stopColor="#9CF3A8" />
+            <stop offset="1" stopColor="#58C96E" />
+          </linearGradient>
+          <linearGradient
+            id="paint3_linear_2620_81262"
+            x1="830.5"
+            y1="1212"
+            x2="830.5"
+            y2="1484"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#01C34B" />
+            <stop offset="1" stopColor="#0799D7" />
+          </linearGradient>
+          <linearGradient
+            id="paint4_linear_2620_81262"
+            x1="830"
+            y1="121"
+            x2="830"
+            y2="1086"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop />
+            <stop offset="0.165" stopColor="#001420" />
+            <stop offset="0.365" stopColor="#002943" />
+            <stop offset="1" stopColor="#008FE8" />
+          </linearGradient>
+          <linearGradient
+            id="paint5_linear_2620_81262"
+            x1="830"
+            y1="1464"
+            x2="830"
+            y2="1714"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#016EDB" />
+            <stop offset="0.555" stopColor="#16B7CD" />
+            <stop offset="1" stopColor="#5CCA69" />
+          </linearGradient>
+        </defs>
+      </svg>
     </div>
   );
 }
-
-const containerStyle = css({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  width: '100%',
-  backgroundColor: '#111827',
-  overflow: 'hidden',
-});
 
 const svgContainerStyle = css({
   width: '100%',
@@ -240,60 +223,6 @@ const svgContainerStyle = css({
 const svgStyle = css({
   width: '100%',
   height: 'auto',
-});
-
-const screenContainerStyle = css({
-  width: '100%',
-  height: '100%',
-  backgroundColor: '#000000',
-  position: 'relative',
-  overflow: 'hidden',
-});
-
-const noiseContainerStyle = css({
-  position: 'absolute',
-  inset: 0,
-  zIndex: 10,
-});
-
-const noiseGridStyle = css({
-  width: '100%',
-  height: '100%',
-  display: 'grid',
-  gridTemplateColumns: 'repeat(12, 1fr)',
-  gridTemplateRows: 'repeat(12, 1fr)',
-});
-
-const noiseCellStyle = css({
-  backgroundColor: '#4ADE80',
-});
-
-const screenContentStyle = css({
-  width: '100%',
-  height: '100%',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  backgroundColor: '#000000',
-});
-
-const screenTextStyle = css({
-  color: '#4ADE80',
-  fontFamily: 'monospace',
-  fontSize: '36px',
-  '@media (min-width: 640px)': {
-    fontSize: '48px',
-  },
-  '@media (min-width: 768px)': {
-    fontSize: '60px',
-  },
-  '@media (min-width: 1024px)': {
-    fontSize: '72px',
-  },
-  '@media (min-width: 1280px)': {
-    fontSize: '96px',
-  },
-  letterSpacing: '2px',
 });
 
 const blurContainerStyle = css({
